add: Added help menu and default prefix events
