+++ .vscode/settings.json
@@ -85,6 +85,7 @@
         "xpamount",
         "xpuser",
         "Yeet",
-        "youdied"
+        "youdied",
+        "ytmp"
     ]
 }
